Extract EAD Tags from CSV Header
As of this commit, the EAD tag names will be extracted from the CSV header
row. Any column named as 'ead:xyztag' will be assumed to be an EAD tag
descriptor, whereby the associated EAD tag name would be 'xyztag'. All
such tags will be packaged into a Python dictionary, and written to the
database.