Implemented move as a copy-and-delete op
In order to be conservative about the transfer process, the move operation
is now implemented as a copy-and-delete. What this means is that when
moving files (as against copying), the source file will be COPIED to the
destination; and once the source and destination checksums are matched,
the source file will be deleted. If, however, there is a checksum
mismatch, then the destination file be deleted. This helps guard against
bit errors during transfers, in which case the source file would be lost
forever in case of a move operation.