Decouple preservation metadata labels from code
The labels to be assigned to preservation metadata items in each record
must be specified outside of the main Python code. This specification is
to be recorded in the labels.json file located in the same directory as
the main script. This file (labels.json) is a JSON formatted file with a
single JSON Object containing key:value pairs. This 'decoupling' allows
users to change the labels assigned to the various metadata items as
appropriate to their usage context.