Convert all paths to absolute paths
Currently the paths specified by the user were used on an as-is basis. But
that situation is not appropriate from a metadata point-of-view, where
relative paths (when supplied by the user) do not convey enough
information. Both source and destination paths will be converted to
absolute paths from now on.