Return db parameters from init_db as a dictionary
Prior to this commit, the init_db() function was returning the db handle
and db collection name packaged into a list. For better readability this
has been modified so that the init_db() function now returns a dictionary
instead of a list. The returned parameters can now be accessed via their
names rather than numeric indices.