Replace ObjectIds with UUIDs for unique file names
The unique file names will now be generated as UUIDs. This was previously
achieved by using MongoDB/BSON ObjectIDs. In order to decouple the choice
of algorithm/specification to generate the unique IDs a new function
'getUniqueID()' has been added. This function provides an abstraction to
the calling code, returning only a unique ID to it and not requiring it to
know which method to generate those IDs.