Add support for resumption of partial transfers
This adds the ability to resume file transfers from where the previous
transfer failed/ended because of some error. Before this change, the only
action taken in case a failure was encountered while transferring the
contents of a directory was to make an entry for it in an errors CSV file.
But now, the script can truly handle an errors CSV file. This is done as
follows: (1) a new field is added to the DB document/record, namely,
'serialNo'; this is the serial number of the file being transferred in the sorted
list for the given source directory. (2) if the destination directory
already exists, the script fetches the largest serial number from the DB
for the corresponding destination directory. (3) the transfers for this
src-dst pair resume from the next serial number in the sorted list of
files in the source directory.