Add basic support for obligation and repeatability
This commit adds basic support for handling the obligation and
repeatability properties of the various metadata entities. Currently the
knowledge of the obligation and repeatability is imbedded in the script,
but in the future it would be decoupled from it. As such that same
knowledge resides in labels.json, but is not being utilized in the script
for simplicity for now.