Refactor code
The global variables are now imported not into the global namespace of the
main script, but in their own namespace called 'g'. So each time a global
variable needs to be accessed, it needs to be accessed as g.<variable
name>.
In the same vein, all the global error code definitions have been moved to
errorcodes.py; and the contents of that module are imported into the main
script into a separate namespace called 'e'. The error codes, therefore,
need to be accessed as e.<error code name>.