Convert lib modules to Python package
All Python modules in the lib/ directory are now packaged into a Python
package called 'metadatautilspkg'. The gzip archive of the current
distribution of the package is contained in the dist/ directory. In order
to install the package to their Python 3.x environment, users must acquire
superuser privilege to execute the INSTALL script on a bash terminal.
Executing INSTALL will unzip and install the package so that the
accession.py script in bin/ can successfully execute.

For developers, it is required that each time they release a new version
of the package, they update the variable called 'VERSION' in the INSTALL
script appropriately before committing the code.